Reader should not try to create an object if such was already created.
In case the eClass attribute is not the first one for an object, the
reader may still be able to create the object based on the type of the
reference in the metamodel. In this case, the reader should not try to
recreate the object again after the eClass attribute has been read,
beacuse other properties will be lost.

Change-Id: I0ca8d060b632525a89711dee73fa21b490bff261